Moved model unit tests into correct folder and namespace.
