+++ src/Umbraco.Tests.UnitTests/Umbraco.Core/Models/ContentScheduleTests.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using NUnit.Framework;
 using Umbraco.Core.Models;
 
-namespace Umbraco.Tests.UnitTests.Umbraco.Infrastructure.Models
+namespace Umbraco.Tests.UnitTests.Umbraco.Core.Models
 {
     [TestFixture]
     public class ContentScheduleTests
