+++ src/Umbraco.Tests.UnitTests/Umbraco.Core/Models/CultureImpactTests.cs
@@ -2,7 +2,7 @@
 using NUnit.Framework;
 using Umbraco.Core.Models;
 
-namespace Umbraco.Tests.UnitTests.Umbraco.Infrastructure.Models
+namespace Umbraco.Tests.UnitTests.Umbraco.Core.Models
 {
     [TestFixture]
     public class CultureImpactTests
