+++ src/Umbraco.Tests.UnitTests/Umbraco.Core/Models/DeepCloneHelperTests.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using NUnit.Framework;
 using Umbraco.Core.Models;
 
-namespace Umbraco.Tests.UnitTests.Umbraco.Infrastructure.Models
+namespace Umbraco.Tests.UnitTests.Umbraco.Core.Models
 {
     [TestFixture]
     public class DeepCloneHelperTests
