+++ src/Umbraco.Tests.UnitTests/Umbraco.Core/Models/DictionaryItemTests.cs
@@ -1,45 +1,45 @@
-﻿using System;
+using System.Linq;
 using Newtonsoft.Json;
 using NUnit.Framework;
 using Umbraco.Core.Models;
 using Umbraco.Tests.Common.Builders;
-using Umbraco.Tests.Common.Builders.Extensions;
 
-namespace Umbraco.Tests.UnitTests.Umbraco.Infrastructure.Models
+namespace Umbraco.Tests.UnitTests.Umbraco.Core.Models
 {
     [TestFixture]
-    public class RelationTypeTests
+    public class DictionaryItemTests
     {
-        private RelationTypeBuilder _builder;
+        private DictionaryItemBuilder _builder = new DictionaryItemBuilder();
 
         [SetUp]
         public void SetUp()
         {
-            _builder = new RelationTypeBuilder();
+            _builder = new DictionaryItemBuilder();
         }
 
         [Test]
         public void Can_Deep_Clone()
         {
             var item = _builder
-                .WithId(1)
-                .WithParentObjectType(Guid.NewGuid())
-                .WithChildObjectType(Guid.NewGuid())
+                .WithRandomTranslations(2)
                 .Build();
 
-            var clone = (RelationType) item.DeepClone();
+            var clone = (DictionaryItem)item.DeepClone();
 
             Assert.AreNotSame(clone, item);
             Assert.AreEqual(clone, item);
-            Assert.AreEqual(clone.Alias, item.Alias);
-            Assert.AreEqual(clone.ChildObjectType, item.ChildObjectType);
-            Assert.AreEqual(clone.ParentObjectType, item.ParentObjectType);
-            Assert.AreEqual(clone.IsBidirectional, item.IsBidirectional);
-            Assert.AreEqual(clone.Id, item.Id);
-            Assert.AreEqual(clone.Key, item.Key);
-            Assert.AreEqual(clone.Name, item.Name);
             Assert.AreEqual(clone.CreateDate, item.CreateDate);
+            Assert.AreEqual(clone.Id, item.Id);
+            Assert.AreEqual(clone.ItemKey, item.ItemKey);
+            Assert.AreEqual(clone.Key, item.Key);
+            Assert.AreEqual(clone.ParentId, item.ParentId);
             Assert.AreEqual(clone.UpdateDate, item.UpdateDate);
+            Assert.AreEqual(clone.Translations.Count(), item.Translations.Count());
+            for (var i = 0; i < item.Translations.Count(); i++)
+            {
+                Assert.AreNotSame(clone.Translations.ElementAt(i), item.Translations.ElementAt(i));
+                Assert.AreEqual(clone.Translations.ElementAt(i), item.Translations.ElementAt(i));
+            }
 
             //This double verifies by reflection
             var allProps = clone.GetType().GetProperties();
@@ -47,12 +47,15 @@ namespace Umbraco.Tests.UnitTests.Umbraco.Infrastructure.Models
             {
                 Assert.AreEqual(propertyInfo.GetValue(clone, null), propertyInfo.GetValue(item, null));
             }
+
         }
 
         [Test]
         public void Can_Serialize_Without_Error()
         {
-            var item = _builder.Build();
+            var item = _builder
+                .WithRandomTranslations(2)
+                .Build();
 
             Assert.DoesNotThrow(() => JsonConvert.SerializeObject(item));
         }
