+++ src/Umbraco.Tests.UnitTests/Umbraco.Core/Models/DictionaryTranslationTests.cs
@@ -6,7 +6,7 @@ using Umbraco.Core.Models;
 using Umbraco.Tests.Common.Builders;
 using Umbraco.Tests.Common.Builders.Extensions;
 
-namespace Umbraco.Tests.UnitTests.Umbraco.Infrastructure.Models
+namespace Umbraco.Tests.UnitTests.Umbraco.Core.Models
 {
     [TestFixture]
     public class DictionaryTranslationTests
