+++ src/Umbraco.Tests.UnitTests/Umbraco.Core/Models/DocumentEntityTests.cs
@@ -4,7 +4,7 @@ using NUnit.Framework;
 using Umbraco.Tests.Common.Builders;
 using Umbraco.Tests.Common.Builders.Extensions;
 
-namespace Umbraco.Tests.UnitTests.Umbraco.Infrastructure.Models
+namespace Umbraco.Tests.UnitTests.Umbraco.Core.Models
 {
     [TestFixture]
     public class DocumentEntityTests
