+++ src/Umbraco.Tests.UnitTests/Umbraco.Core/Models/MacroTests.cs
@@ -6,7 +6,7 @@ using Umbraco.Core.Models.Entities;
 using Umbraco.Tests.Common.Builders;
 using Umbraco.Tests.Common.Builders.Extensions;
 
-namespace Umbraco.Tests.UnitTests.Umbraco.Infrastructure.Models
+namespace Umbraco.Tests.UnitTests.Umbraco.Core.Models
 {
     [TestFixture]
     public class MacroTests
